Safari-609.1.11.1/LayoutTests/fast/misc/{testProcessIncomingSyncMessagesWhenWaitingForUnboundedReply.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<script>
description("Test for the SendSyncOption::ForceDispatchWhenDestinationIsWaitingForUnboundedSyncReply sendSync flag");

shouldBeTrue("internals.testProcessIncomingSyncMessagesWhenWaitingForSyncReply()");
</script>
</body>
</html>
