Safari-609.1.11.1/LayoutTests/fast/misc/{testProcessIncomingSyncMessagesWhenWaitingForUnboundedReply-expected.txt

Test for the SendSyncOption::ForceDispatchWhenDestinationIsWaitingForUnboundedSyncReply sendSync flag

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.testProcessIncomingSyncMessagesWhenWaitingForSyncReply() is true
PASS successfullyParsed is true

TEST COMPLETE
